Desktop/cognitivelinguistics/requirements.txt
requests
beautifulsoup4
pandas
newspaper3k
dbt-duckdb
